+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix a bug: the subscription page is not available
 - Fix a bug: users promoted to the administrator role are not in the admin group
 - Fix a bug: menu separators are not visible
+- Fix a bug: prevent VersionCheckWorker from polluting the sidekiq stack in development
 - [TODO DEPLOY] `rails fablab:maintenance:rebuild_stylesheet`
 
 ## v4.4.0 2020 May 12